feat(build): Update Makefile to migrate build system to Vitis flow
This commit replaces the original Makefile with a new version built from scratch for Vitis to enable targeting of Alveo data center accelerator cards.

Key Changes:

- Replaces manual script-based HLS and Vivado flow with streamlined Vitis v++ compiler.

- Introduces standard Vitis build targets: 'sw_emu', 'hw_emu' and 'hw'.

The previous Makefile designed for a Kria embedded flow is now deprecated and has been removed and this new structure is made for easier future kernel development on Alveo.